Create Notes from Tony
General comments: [1] Your Rmd file did not knit until I removed the install() function calls. Please do not include those calls in an Rmd file as those will typically throw errors and remember to do a test rendering of your Rmd file to make sure it renders correctly and includes the information/output expected (see below) [2] Please be sure that your code PRINTS OUT the results of analyses/plots I ask you to do. Here, for Steps 1, 3, 8, nothing is printed to the console. You do the wrangling and analyses I ask you to do, but don't print those results... E.g., print out `s` to the console in Step 1. Print out `beta1` and `beta0` in Step 3. Print out `groups` in Step 5. Step 5: your lm() here doesn't work as you expect... you are running it with data = groups, but groups at this point is a just a table with beta1 and beta0 for catarrhines, platyrrhines, and strepsirrhines. Step 6: you don't actually calculate the SE of CI for beta1 by hand as your model object (m) is not really the model you think it is (see Step 5 comments above). The fact that your code returns NaN for many of the intermediate variables you calculate here should give you a hint that something's up. Check out the posted solution for more details on how to proceed here. Step 7: works fine! Step 8: t.test is not the way to judge this here... the p value returned for this test may be larger than 0.05, but what are you looking at here? whether the bootstrapped value for the slope coefficient is different than the actual value. Such a test really doesn't make sense to do as bootstrapped samples from the same initial sample are unlikely to generate a distribution very different from the original sample. Here, all you need to do is bootstrap to generate an estimated CI and look at whether that interval included zero or not (does it?)